Added uncertainty input from separate excel sheet
To structure the process of managing inputs, it is now possible to speciy input data in a separate excel sheet. Currently only a small example implemented (see "test_input_read.py") only with uncertainties. Next steps is to add all paramter types and same thing for policies so both polices and uncertainties can be specified in a separate file.